Override shortcuts in SyntaxLineEdit.
Add an ::event member function to SyntaxLineEdit that looks for
ShortcutOverride events and accepts them. This keeps them from being
interpreted as shortcuts in the main window.

Note that the current code accepts everything, and that we should probably
restrict this to AltGr only.

Bug: 12270
Change-Id: I01765ce2447d220a102d97fcbbe47579341ce075
Reviewed-on: https://code.wireshark.org/review/14570
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Gerald Combs <gerald@wireshark.org>