Fix byte pane not showing up in Packet dialog
When a packet dialog is being constructed, the layout is apparently not
fixed yet and the byte view is technically not visible. Fix the hidden
byte pane by not hiding it when it is not (yet) visible.

Bug: 11760
Change-Id: I0494fa16a5ed89ff31f934ba682a6bb884cc0e2e
Reviewed-on: https://code.wireshark.org/review/14713
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
(cherry picked from commit 4de738f5a720012afde1c9e38b1cbf1131fdee34)
Reviewed-on: https://code.wireshark.org/review/14722