Clean up some stuff.
Have bin2hex() wmem_allocate the buffer, so it can be used the same way
that tvb_get_string_enc() is used.

Don't bother checking whether NIBBLE_2_ASCHEX() returns an ASCII hex
digit character or not - it returns either a value in the ASCII range of
'0' through '9' or in the range 'A' through 'F', all of which are ASCII
hex digits.

Fix get_bit() to set *length to 0 if the string we're returning is
empty.

Change-Id: Id331cfd0ab34d45892f98d228dc793a1e93d84e5
Reviewed-on: https://code.wireshark.org/review/14717
Reviewed-by: Guy Harris <guy@alum.mit.edu>