Qt: prevent input to the capture filter field during initialization
Mark the capture filter field as readonly until initialization is
completed. This avoids useless entries in the recent capture filters
list and fixes a use-after-free when typing during startup.

(This use-after free occurred when a new item is added to the recent
filters list which invalidates the previously retrieved list).

Change-Id: Ifae9444986436a60781bf5f83a6e650501ccc4d5
Reviewed-on: https://code.wireshark.org/review/14740
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>