Fix packet length handling.
Treat the packet length as unsigned - it shouldn't be negative in the
file.  If it is, that'll probably cause the sscanf to fail, so we'll
report the file as bad.

Check it against WTAP_MAX_PACKET_SIZE to make sure we don't try to
allocate a huge amount of memory, just as we do in other file readers.

Use the now-validated packet size as the length in
ws_buffer_assure_space(), so we are certain to have enough space, and
don't allocate too much space.

Bug: 12394
Change-Id: If0b79376fc6fe2f49c7b7d7796dee7b7683485cb
Reviewed-on: https://code.wireshark.org/review/15171
Reviewed-by: Guy Harris <guy@alum.mit.edu>