Don't assume packets will be no larger than ISERIES_MAX_PACKET_LEN.
We don't check against it.  Insteead, use phdr->caplen as the buffer
size; that's based on the number of hex digits we've found.

While we're at it, also get rid of ISERIES_PKT_ALLOC_SIZE - it makes it
less obvious that it's based on the packet length from the packet
header.

Change-Id: I7a3fac130adc75aa0a4b426cb9b5c9c6279e1c88
Reviewed-on: https://code.wireshark.org/review/15238
Reviewed-by: Guy Harris <guy@alum.mit.edu>