s7comm: block/file up-/download redesign
Redesigned the block/file up- and download, using new informations from
PI-Services.
Changed the block-info items, as it's in general a string, which can be
further dissected if it's a PLC program block.
Other types of Siemens automation hardware may use this as a simple
filename for filetransfer.
Changed the display-style for block name/number in info-column, to be more
compact.

Change-Id: Ie7c3a285d25897f8516ed8827041a5dbdf80679e
Reviewed-on: https://code.wireshark.org/review/15275
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>