Fix dissection of JXTA over UDP.
Prior to this change all JXTA-over-UDP frames throw an exception after setting
COL_PROTOCOL.

The offset used when adding the protocol to the tree had already been
incremented to point to the end of the TVB; use tree_offset (which is the
current offset) instead.

This code has been here since r17282 (2006) so presumably the change is that
proto_tree_add_protocol_format() started throwing an exception when offset
points to the end of the TVB sometime between then and now.

Change-Id: Iadbcada29cd235e8abe5662b825a2d97ff6f7308
Reviewed-on: https://code.wireshark.org/review/15560
Petri-Dish: Jeff Morriss <jeff.morriss.ws@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>