Restore a preference column type if fields are empty.
The column type is automatically changed to Custom if user clicks
on the "Fields" or "Field Occurrences" column.  This can be
accidental and cause the user grief if he doesn't realize it.
If there is no text in the field, restore the type to its
original value.

# Conflicts:
#	ui/qt/column_preferences_frame.cpp

Bug: 12465
Change-Id: I0453f43be13077aba58cca61011eeff9ca9b83b2
Reviewed-on: https://code.wireshark.org/review/15600
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>
Reviewed-on: https://code.wireshark.org/review/15658
Reviewed-by: Michael Mann <mmann78@netscape.net>