Make the modelines more like other ones.
Thou shalt not ever use modelines that put the tab-character tab spots
anywhere other than every 8 characters; that's where Ken and Dennis put
them, and that's where they belong.  Use whatever indentation you want,
including 4-space indentation, but do *not* try to arrange that a tab
character moves to the next 4-character boundary, because, in a lot of
UN*X software, it doesn't.

(Yes, this means that Xcode's default is wrong.  It *is* wrong,
especially given that it's an IDE for a UNIX.)

Change-Id: I308745cdeef35b7c91ea493da6487baadc357f58
Reviewed-on: https://code.wireshark.org/review/15673
Reviewed-by: Guy Harris <guy@alum.mit.edu>