Don't assume all IDBs are available aftre we open the file.
IDBs can occur anywhere in the file, so if we see an interface ID bigger
than the number of IDBs we've see, re-fetch the interface information,
update the IDB count, and grow the packet count array as necessary.

Get the information strings for interfaces after reading the entire
file; we don't need them until then.

Change-Id: I54dfbf072a5aa53369318d1a04bb9d11cc40ec4a
Reviewed-on: https://code.wireshark.org/review/15799
Reviewed-by: Guy Harris <guy@alum.mit.edu>