cmake: add PIDL dissector generation support
The dissectors list is generated based on regex magic on pidl/Makefile.
A dssetup.cnf file is added because all other protocols already have
this file. The srvsvc and wkssvc protocols cannot be built for some
reason, so just disable it (this mirrors the original Makefile).

Add frsrpc and samr to PIDL_DISSECTOR_SRC list which was missing
compared to autotools.

Tested with an out-of-tree and in-tree build with cmake (make and ninja):

    cmake -GNinja && ninja -v pidl-dissectors
    touch epan/dissectors/pidl/dfs/dfs.idl && ninja generate_dissector-dcerpc-dfs
    cmake . && make -C epan/dissectors/pidl generate_dissector-dcerpc-dfs

Change-Id: Id8bf34e76c20053a8a393024560d90a048473e03
Reviewed-on: https://code.wireshark.org/review/15962
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>