NHRP: Add option for source address in authentication extension.
It seems that not all Cisco IOS implementations adhere to RFC2332
with respect to having the source address after the SPI in the
authentication extension. This change adds a preference to suppress
the interpretation of the bytes following the SPI as the source
address. The default is to adhere to the RFC.

Bug: 12569
Change-Id: I00d3c1b90ace54f16f0fe4704e6127a5c4881c82
Reviewed-on: https://code.wireshark.org/review/16231
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Reviewed-by: Michael Mann <mmann78@netscape.net>