packet-ehdlc: Update with results from better understanding the protocol
The Ericsson HDLC format inside L2TP is not publicly documented, so when
I originally wrote the code, I was basically just looking at hex-dumps
of protocol traces and trying to ge an initial understanding. By now,
that understanding has significantly matured, so let's update the
dissector accordingly.

The major changes are regarding the understanding that there is not a
message type, but in fact a compressed SAPI and compressed TEI field.

Based on this, we can discriminate on whether there will be a HDLC/LAPD
header to follow, or whether there's actual user data (TRAU frame
replacement) following.

Change-Id: Ic29d2d295678c4dc5f93cd8404b2e832f4532daf
Reviewed-on: https://code.wireshark.org/review/16508
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>